requirements.txt
dash==2.11.1
dash_bootstrap_components==1.4.2
dash_mantine_components==0.12.1
pandas==2.0.3
Pillow==10.4.0
plotly==5.15.0
